pyproject.toml
[project]
name = "garage8-agent"
version = "0.1.0"
description = "Garage8 AI Agent"
authors = [
    {name = "Slava Golyshevskii",email = "sgolyshevskii@gmail.com"}
]
license = {text = "LICENSE"}
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-telegram-bot (>=21.10,<22.0)",
    "pydantic-ai[logfire] (>=0.0.23,<0.0.24)",
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sshtunnel (>=0.4.0,<0.5.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "halo (>=0.0.31,<0.0.32)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "einops (>=0.8.1,<0.9.0)",
    "qdrant-client (>=1.13.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
